Refactoring: Insert new InputTags later in the process when everything is inserted into the db.
Note: Better would be to have this transactional by letting SearchInputWithRules.update insert new tags and link them accordingly in one go. But the side effect of ending up with unused tags in case of errors seems unproblematic since the user would likely retry the import so that the tags end up in the db eventually.